client: Add new event when WebApp's ui is loaded
Add Zarafa.onUIReady which fires when the WebApp has drawn the Main
viewport and has loaded the hierarchy. This makes it possible for
plugins or Webapp code to register to this event instead on
Zarafa.onReady which leads to an login performance improvement.

The container event 'webapploaded' cannot be used, since the container
is only avaliable when WebApp has fired 'fireReady'.

References: KW-1367