client: Load GAB hierarchy when WebApp UI is rendered
Instead of loading the GAB hierarchy when the WebApp is still building
up the WebApp UI and loading the hiearchy, load it after the UI has
been rendered. This improves login performance, since the user does not
need the GAB hierarchy directly after being logged in, in most cases.

References: KW-1368